TEST_v1.3 This project has both the uBlaze + DDR in AXI configuration and the VHDL FMS + DDR in user interface configuration. Both of them are failing because of the bug in Vivado. I am trying to move the project in Vivado 2018.1
In the meantime Terry K. jumper the uCOE signal from the test point at fpga and tie it to a pin on the fpga debug 10 pin header.
So next commit will comment all the DDR part and will route the uCRD signal to one pin of the debug header to test the write and read with the uController.